replace libvips-cpp with custom-built libvips
looks like the static libvips-cpp distribution i was using stripped a few of the symbols
from the libvips c api :( i think because it was treating the c api as a library dependency
of the cpp api. but our jna is only using the c api, so i rebuilt the static lib to exclude
the cpp api, making the c api the primary target, so the symbols should be kept now.

if this works then i think the best approach is probably to fork that build script into a
guardian repo, and then we can rebuild as needed, though ill also raise an issue in their
repo to see if they think its a problem and need to tweak the symbol stripping config